requirements.txt
asgiref
Django
django-ckeditor
django-crispy-forms
django-import-export
django-js-asset
Pillow
gunicorn
sqlparse
